Changed the action types, streamlining the process.
All actions now the use logic of selector nodes, the BT_ActionNode.cpp is a bit redundant as of the moment, for code refactoring reasons. In short, because I don't use any node types nor after few days of planning could not find a reason to include new types for the short-term future, decided to use one class (Selector nodes) for simplicity.

Also added a new node state, Skip, where we can Skip the entire action but keep executing the children, for example the GetBall action where it should no longer execute if we have the ball.

The 4 types:

- Simple action, while it is a selector node, it behaves exactly like an action node.
- Simple Chain Action - executes all children without picking a specific one or checking conditions. This sort of behaves like a mix between parallel/decorator actions. The goal of these to execute action before others while the outcome not being a requirement for the children. One use case now is determining dribble direction and dribbling, where determining dribble dir cannot 'fail' (in the sense of it will always produce something) and should execute Dribble anyway. It will execute all children unless all succeed or all fail.
- Selectors: Just some logical changes. Runs all children without condition checking unless one fails. At the moment, I don't have a use case for this and it exists because it is a classic BT node type.
- Picker: no logical changes, it picks one specific child based on PickNextAction function. If it can pick an action, it will execute that child, otherwise it will run its own action.

Also added few new actions that are WIP as I am working on my first 1 v 1 interaction, where the AI needs to overcome another AI via dribbling.

Also for that reason changed the logic of ball ownership, making a new 'controlling' bool where the player can have the ball but not control it, for example when pushing the ball for dribbling. This will be useful when a defender trying to take the ball away while knowing that itself does not have the ball.

+ Minor refactoring here and there, forgot :(. Must admit, should have pushed sooner.

Now that the GOBP aspect is functioning reliably, I can probably fully focus on the 1 v 1 interaction. There are still few caveats that needs to be ironed out, however, they are cases that currently do not affect existing actions, so I will handle that afterward. They have been marked with comments or in my own notes.